# packages/billing-engine/openapi.yaml
openapi: 3.0.1
info:
  title: MARP Billing Engine API
  version: "0.1.0"
  description: API for subscription lifecycle and billing operations
servers:
  - url: http://localhost:3100
paths:
  /marp/subscription/create:
    post:
      summary: Create a subscription (signup)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscription'
      responses:
        '200':
          description: ledger entry or scheduling note
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LedgerEntry'
  /marp/subscription/renew:
    post:
      summary: Renew a subscription (explicit)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenewSubscription'
      responses:
        '200':
          description: ledger entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LedgerEntry'
  /marp/subscription/upgrade:
    post:
      summary: Upgrade subscription (prorated)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpgradeSubscription'
      responses:
        '200':
          description: ledger entry or scheduling note
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/LedgerEntry'
                  - $ref: '#/components/schemas/ScheduleNote'
  /marp/subscription/cancel:
    post:
      summary: Cancel subscription (effective period end)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelSubscription'
      responses:
        '200':
          description: cancellation scheduled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancellationNote'
components:
  schemas:
    CreateSubscription:
      type: object
      required: [accountId, walletId, planId, price]
      properties:
        accountId:
          type: string
          description: Account identifier
        walletId:
          type: string
        planId:
          type: string
        price:
          type: number
          minimum: 0
        region:
          type: string
        at:
          type: string
          format: date-time
        idempotencyKey:
          type: string
        autoRenew:
          type: boolean
    RenewSubscription:
      type: object
      required: [accountId]
      properties:
        accountId:
          type: string
        at:
          type: string
          format: date-time
        idempotencyKey:
          type: string
    UpgradeSubscription:
      type: object
      required: [accountId, walletId, newPlanId, newPrice]
      properties:
        accountId:
          type: string
        walletId:
          type: string
        newPlanId:
          type: string
        newPrice:
          type: number
          minimum: 0
        at:
          type: string
          format: date-time
        idempotencyKey:
          type: string
    CancelSubscription:
      type: object
      required: [accountId]
      properties:
        accountId:
          type: string
        effectiveImmediately:
          type: boolean
          description: Whether to cancel immediately (not allowed by policy)
    LedgerEntry:
      type: object
      properties:
        entryId:
          type: string
        walletId:
          type: string
        accountId:
          type: string
        amount:
          type: number
        prevHash:
          type: string
          nullable: true
        entryHash:
          type: string
        balanceAfter:
          type: number
        ts:
          type: string
          format: date-time
    ScheduleNote:
      type: object
      properties:
        scheduled:
          type: boolean
        effectiveFrom:
          type: string
          format: date-time
        note:
          type: string
    CancellationNote:
      type: object
      properties:
        cancelled:
          type: boolean
        effectiveFrom:
          type: string
          format: date-time
        note:
          type: string